pyproject.toml
[project]
name = "donation-api"
version = "0.1.0"
description = "Rest API for Donation App."
authors = [
    { name = "Kumar Swaraj", email = "123369109+kumar-swaraj@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "django-import-export (>=4.4.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "stripe (>=14.3.0,<15.0.0)",
]

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


[tool.ruff]
line-length = 88
indent-width = 4
exclude = [".git", ".venv", "__pycache__", "**/migrations/**", "manage.py"]

target-version = "py314"

[tool.ruff.lint]
select = [
    "B",
    "C",
    "DTZ",
    "DJ",
    "EM",
    "FIX",
    "ICN",
    "T20",
    "SLF",
    "SIM",
    "TID",
    "I",
    "C90",
    "N",
    "PERF",
    "E",
    "W",
    "F",
    "UP",
]
# ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
docstring-code-format = false


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-ra -q"
